Added error handling and restructured code
