+++ websites/__init__.py
@@ -1,0 +1,5 @@
+from .task_one import *
+from .task_two import *
+from .task_three import *
+from .task_four import *
+from .utils import *
